backend/pyproject.toml
[project]
name = "tsingai-lens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "markdown>=3.6",
    "networkx>=3.3",
    "openai>=1.40.0",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.24.4",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
    "yake>=0.4.8",
]
